Cast to long for long long output (%lld not portable)
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@4516 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/collections/OutputStream.m
@@ -135,7 +135,7 @@ PHASE(Using)
 {
   char buf[DSIZE (long long) + 1];
 
-  sprintf (buf, "%lld", lnglng);
+  sprintf (buf, "%ld", (long) lnglng);
   [self catC: buf];
 }
 
@@ -143,7 +143,7 @@ PHASE(Using)
 {
   char buf[DSIZE (unsigned long long) + 1];
 
-  sprintf (buf, "%llu", ulnglng);
+  sprintf (buf, "%lu", (unsigned long) ulnglng);
   [self catC: buf];
 }
